Added Homepage + not-implemented User Options
Files in this upload include the Home Page screen for the app. HomePage.java has been made the new main activity launcher, so it should be the first thing to see when running app tests on Android Studio.

Note: The MyArchivePage.xml file has a working "Return to Home" button, but the AppSettings.xml and MyProfile.xml don't. I'm still currently figuring out why this is happening at the moment.